--- a/.github/workflows/manifold.yml
+++ b/.github/workflows/manifold.yml
@@ -26,7 +26,7 @@ jobs:
     - name: Install dependencies
       run: |
         apt-get -y update
-        DEBIAN_FRONTEND=noninteractive apt install -y cmake libglm-dev libassimp-dev libboost-graph-dev
+        DEBIAN_FRONTEND=noninteractive apt install -y cmake libglm-dev libassimp-dev libboost-graph-dev libcgal-dev
     - name: Build CUDA
       run: |
         mkdir buildCUDA