--- a/tools/CMakeLists.txt
+++ b/tools/CMakeLists.txt
@@ -15,6 +15,12 @@ target_compile_features(perfTest PUBLIC cxx_std_14)
 add_executable(perfTestCGAL perf_test_cgal.cpp)
 find_package(CGAL REQUIRED COMPONENTS Core)
 target_compile_definitions(perfTestCGAL PRIVATE CGAL_USE_GMPXX)
+# target_compile_definitions(perfTestCGAL PRIVATE CGAL_DEBUG)
+target_link_libraries(perfTestCGAL manifold CGAL::CGAL CGAL::CGAL_Core)
+
+target_compile_options(perfTestCGAL PRIVATE ${MANIFOLD_FLAGS})
+target_compile_features(perfTestCGAL PUBLIC cxx_std_14)
+
 # add_executable(playground playground.cpp)
 # target_link_libraries(playground manifold meshIO samples)