First off: amazing work, I'm super excited about Manifold and hope we can use it in OpenSCAD and beyond!

I've ported perf_test to use CGAL's corefinement after reading [your post](https://elalish.blogspot.com/2022/03/manifold-performance.html),
which I believe tests OpenSCAD more than it does CGAL itself.

I haven't had a chance to test with CUDA hardware or with OpenMP (which I expect are gonna leave CGAL in the dust, since I don't think corefinement is one of their parallelizable operations), but thought it worth doing a mono-threaded comparison as fairly as possible.

To that effect I've used the `CGAL::Epick` [kernel](https://doc.cgal.org/latest/Kernel_d/group__PkgKernelDKernels.html) (rather than the exact, [GMP](https://gmplib.org/) rationals-based `CGAL::Epeck` we use in OpenSCAD). `Epick` is only marginally faster than `Epeck` here anyway, but note that both provide exact predicates, `Epick` is only "inexact" in that it's built from double values (much like Manifold, AFAICT).

Haven't tested all the operations but hopefully you can vary the test cases using this PR as a starting point.

Some take-aways from only testing union and difference (again, only CPU-monothreaded)
- Manifold's union is up to 2.3x faster than corefinement for meshes >32k faces, but...
- Corefinement union 2-3x faster for small meshes <=2k faces
- Corefinement difference is faster at all mesh sizes I've tested

(see [raw results here](https://docs.google.com/spreadsheets/d/13ze7G1JTXOtgKhyWsBr890G-kdcGM-aG-fp6QcWManw/edit#gid=0), took best of 3 runs)

Tested with:

```bash
git clone https://github.com/ochafik/manifold.git --single-branch -b perf_test_cgal
cd manifold
docker run -v $PWD:/src:z -it nvidia/cuda:11.1-devel-ubuntu20.04 

# Then:
apt-get -y update
DEBIAN_FRONTEND=noninteractive apt install -y cmake libglm-dev libassimp-dev libboost-graph-dev libcgal-dev

cd /src
mkdir -p buildCPP && cd buildCPP
cmake -DCMAKE_BUILD_TYPE=Release -DMANIFOLD_USE_CPP=ON .. && make
tools/perfTest
tools/perfTestCGAL
```

@sloriot FYI more microbenchmarking action :-)