@elalish 2022-03-08
Thanks for the raw results; super interesting! I'm curious why OpenSCAD is an order of magnitude slower than CGAL here... 

I'm stoked that mine is performing pretty nearly as well; I haven't spent much time on performance optimization and I have the feeling there is some low-hanging fruit.