@ochafik 2022-03-08
> I'm curious why OpenSCAD is an order of magnitude slower than CGAL here...

I'm curious too haha! In UI mode (as opposed CLI) I haven't implemented new rendering logic so the results are converted to a mesh (PolySet) the existing renderer can process, but other things could be going wrong, I'll investigate :-)