@ochafik 2022-03-08
> I'm stoked that mine is performing pretty nearly as well as this new CGAL; I haven't spent much time on performance optimization and I have the feeling there is some low-hanging fruit.

Super impressive indeed 👏, and exciting times ahead!