@ochafik 2022-03-09
> saves me having to learn CGAL. Are you going to do the Sponge next?

@elalish when I have a bit more time I'll try and repurpose some CLI I wrote that reads .off files and does union/diff/inter with different kernels and now different libraries, so you can do all the tests you want ;-)