@pca006132 2022-05-06
It seems that quite a bit of time is spent on initializing the vectors, e.g. halfedge vectors, that we will overwrite them later anyway. [Here](https://github.com/pca006132/manifold/commit/f181c0d0e53fde387c3833c6843de53d93bf3c65) is a patch that avoids such initialization (a bit ugly and unsafe, not sure if I should open a PR now) that improves the performance of single threaded manifold operations on small meshes.

Before this patch (single threaded perfTest):
```
nTri = 512, time = 0.0107632 sec
nTri = 2048, time = 0.0160586 sec
nTri = 8192, time = 0.028829 sec
nTri = 32768, time = 0.0673084 sec
nTri = 131072, time = 0.242107 sec
nTri = 524288, time = 0.941614 sec
nTri = 2097152, time = 4.21132 sec
nTri = 8388608, time = 17.4662 sec
```

After this patch:
```
nTri = 512, time = 0.00140315 sec
nTri = 2048, time = 0.00383832 sec
nTri = 8192, time = 0.0140288 sec
nTri = 32768, time = 0.0530416 sec
nTri = 131072, time = 0.236449 sec
nTri = 524288, time = 0.917415 sec
nTri = 2097152, time = 4.02338 sec
nTri = 8388608, time = 17.8529 sec
```

CGAL for comparison:
```
nTri = 512, time = 0.00349265 sec
nTri = 2048, time = 0.00595918 sec
nTri = 8192, time = 0.0176475 sec
nTri = 32768, time = 0.0586701 sec
nTri = 131072, time = 0.277333 sec
nTri = 524288, time = 1.20353 sec
nTri = 2097152, time = 5.51911 sec
nTri = 8388608, time = 25.991 sec
```

Sadly this patch cannot improve the performance for OpenMP and CUDA backend. I guess there are still quite a lot of things that can be optimized better.