@elalish 2022-05-06
Feel free to open a PR whenever you'd like feedback on it. I definitely like the idea of performance optimizations. Part of the reason I'd like to move to C++17 parallel algorithms is that it would allow me to remove the entire `VecDH` class, as that would then be handled by the compiler internally.